+++ src/pythia/src/types/mod.rs
@@ -7,3 +7,4 @@ pub mod state;
 pub mod subscription;
 pub mod whitelist;
 pub mod withdraw;
+pub mod timer;