+++ CHANGELOG.md
@@ -15,6 +15,9 @@ All notable changes to this project will be documented in this file.
 - A method to update a subscription
 - The update_chain_fee_and_symbol method
 - Protection for insufficient funds
+- A method to transfer all funds from chains from controllers
+- A method to remove a subscription
+- A method to stop the main timer
 
 ### Fixed
 - Refresh README.md